Conflicts:
	CorePlayer/src/main/flex/com/seesaw/player/SeeSawPlayer.as

+++ BatchEventServices/src/main/flex/com/seesaw/player/batchEventService/BatchEventService.as
@@ -15,15 +15,14 @@ import com.seesaw.player.namespaces.contentinfo;
 import com.seesaw.player.services.ResumeService;
 
 import flash.events.TimerEvent;
+import flash.external.ExternalInterface;
 import flash.utils.Timer;
 
 import org.as3commons.logging.ILogger;
 import org.as3commons.logging.LoggerFactory;
 import org.osmf.elements.ProxyElement;
 import org.osmf.events.BufferEvent;
-import org.osmf.events.DisplayObjectEvent;
 import org.osmf.events.DynamicStreamEvent;
-import org.osmf.events.LoadEvent;
 import org.osmf.events.MediaElementEvent;
 import org.osmf.events.MetadataEvent;
 import org.osmf.events.PlayEvent;
@@ -34,7 +33,6 @@ import org.osmf.media.MediaElement;
 import org.osmf.metadata.Metadata;
 import org.osmf.net.StreamingURLResource;
 import org.osmf.traits.BufferTrait;
-import org.osmf.traits.DisplayObjectTrait;
 import org.osmf.traits.DynamicStreamTrait;
 import org.osmf.traits.LoadTrait;
 import org.osmf.traits.MediaTraitType;
@@ -63,13 +61,13 @@ public class BatchEventService extends ProxyElement {
 
     private var seeking:Boolean;
 
-    private var transactionItemId:uint;
+    private var transactionItemId:int;
     private var serverTimeStamp:uint;
-    private var mainAssetId:uint;
-    private var sectionCount:uint;
-    private var programmeId:uint;
-    private var userId:uint;
-    private var anonymousUserId:uint;
+    private var mainAssetId:int;
+    private var sectionCount:int;
+    private var programmeId:int;
+    private var userId:int;
+    private var anonymousUserId:int;
 
     private var contentViewingSequenceNumber:int = 0;
     private var currentAdBreakSequenceNumber:int = 0;
@@ -99,14 +97,19 @@ public class BatchEventService extends ProxyElement {
     private var viewEvent:ViewEvent;
     private var userEvent:UserEvent;
     private var availabilityType:String;
+    private var adUrlResource:String;
+    private var oldUserEventId:int = 0;
 
-    public function BatchEventService(proxiedElement:MediaElement = null) {
-        var provider:ObjectProvider = ObjectProvider.getInstance();
-        resumeService = provider.getObject(ResumeService);
-        if (resumeService == null) {
-            throw ArgumentError("no resume service implementation provided");
-        }
-    }
+       public function BatchEventService(proxiedElement:MediaElement = null) {
+           var provider:ObjectProvider = ObjectProvider.getInstance();
+           resumeService = provider.getObject(ResumeService);
+           if (resumeService == null) {
+               throw ArgumentError("no resume service implementation provided");
+           }
+           if (ExternalInterface.available) {
+               ExternalInterface.addCallback("exitPlayerWindow", exitEvent);  // fire the exit event hooked into the widow.onUnLoad we currently use...
+           }
+       }
 
     public override function set proxiedElement(proxiedElement:MediaElement):void {
         if (proxiedElement) {
@@ -125,7 +128,7 @@ public class BatchEventService extends ProxyElement {
             cumulativeDurationCount = 0;
             cumulativeDurationMonitor = new Timer(CUMULATIVE_DURATION_MONITOR_TIMER_DELAY_INTERVAL, 0);
             cumulativeDurationMonitor.addEventListener(TimerEvent.TIMER, incrementCumulativeDurationCounter);
-            cumulativeDurationMonitor.start();
+
 
             cumulativeDurationFlushTimer = new Timer(CUMULATIVE_DURATION_FLUSH_DELAY_INTERVAL, 0);
             cumulativeDurationFlushTimer.addEventListener(TimerEvent.TIMER, onTimerTick);
@@ -133,14 +136,14 @@ public class BatchEventService extends ProxyElement {
 
             playerMetadata = proxiedElement.resource.getMetadataValue(PlayerConstants.METADATA_NAMESPACE) as Metadata;
             playerMetadata.addEventListener(MetadataEvent.VALUE_CHANGE, playerMetaChanged);
-             playerMetadata.addEventListener(MetadataEvent.VALUE_ADD, playerMetaChanged);
+            playerMetadata.addEventListener(MetadataEvent.VALUE_ADD, playerMetaChanged);
             playerMetadata.addEventListener(MediaElementEvent.METADATA_ADD, playerMetaChanged);
 
 
             if (playerMetadata) {
                 transactionItemId = playerMetadata.getValue("videoInfo").transactionItemId;
                 serverTimeStamp = playerMetadata.getValue("videoInfo").serverTimeStamp;
-                mainAssetId = playerMetadata.getValue("videoInfo").mainAssetID;
+                /// mainAssetId = playerMetadata.getValue("videoInfo").mainAssetID; this is nolongerNeeded
                 batchEventURL = playerMetadata.getValue("contentInfo").batchEventService;
                 cumulativeDurationURL = playerMetadata.getValue("contentInfo").cumulativeDurationService;
                 sectionCount = playerMetadata.getValue("videoInfo").sectionCount;
@@ -150,7 +153,7 @@ public class BatchEventService extends ProxyElement {
                 adMode = playerMetadata.getValue("contentInfo").adMode;
                 availabilityType = playerMetadata.getValue("videoInfo").availabilityType;
 
-                viewEvent = new ViewEvent(transactionItemId, serverTimeStamp, sectionCount, mainAssetId, userId, anonymousUserId);
+
 
                 var number:Number = resumeService.getResumeCookie();
                 if (number == 0) {
@@ -158,7 +161,8 @@ public class BatchEventService extends ProxyElement {
                 } else {
                     userEvent = buildAndReturnUserEvent(UserEventTypes.AUTO_RESUME);
                 }
-                if (adMode != AdMetadata.LR_AD_TYPE || AdMetadata.AUDITUDE_AD_TYPE) {
+                if (adMode != AdMetadata.LR_AD_TYPE && adMode != AdMetadata.AUDITUDE_AD_TYPE) {
+                    viewEvent = new ViewEvent(transactionItemId, serverTimeStamp, sectionCount, mainAssetId, userId, anonymousUserId);
                     eventsManager = new EventsManagerImpl(viewEvent, availabilityType, batchEventURL, cumulativeDurationURL);
                     eventsManager.addUserEvent(userEvent);
                     eventsManager.flushAll();
@@ -169,11 +173,27 @@ public class BatchEventService extends ProxyElement {
 
     private function playerMetaChanged(event:MetadataEvent):void {
         //TODO  lIVERAIL METDATA CHANGE AGAINST THE SECTIONCOUNT BEFORE THIS IS FIRED......
-        if (event.key == AdMetadata.AD_STATE) {
-            eventsManager = new EventsManagerImpl(viewEvent, availabilityType, batchEventURL, cumulativeDurationURL);
-            eventsManager.addUserEvent(userEvent);
-            eventsManager.flushAll();
+        if (adMode == AdMetadata.LR_AD_TYPE || AdMetadata.AUDITUDE_AD_TYPE) {
+            if (event.key == AdMetadata.SECTION_COUNT) {
+                sectionCount = evaluateNewSectionCount(event.value);
+                viewEvent = new ViewEvent(transactionItemId, serverTimeStamp, sectionCount, mainAssetId, userId, anonymousUserId);
+                eventsManager = new EventsManagerImpl(viewEvent, availabilityType, batchEventURL, cumulativeDurationURL);
+                eventsManager.addUserEvent(userEvent);
+                eventsManager.flushAll();
+            }
         }
+    }
+
+    private function evaluateNewSectionCount(value:int):int {
+        var newSectionCount:int;
+            /// SMILResource should only have one asset in the event of liverail or auditude and we ALWAYS presume there is a preRoll
+            // ELSE this rule will fail..
+             if(value == 1 && sectionCount == 1){
+                newSectionCount =  value + sectionCount;
+             }else{
+               newSectionCount = value*2;
+             }
+        return  newSectionCount;
     }
 
 
@@ -207,10 +227,7 @@ public class BatchEventService extends ProxyElement {
             adMetadata.addEventListener(MetadataEvent.VALUE_CHANGE, onAdsMetaDataChange);
 
         } else if (event.namespaceURL == PlayerConstants.SMIL_METADATA_NS) {
-            // TODO this seems to be the only way currently to detect main content is being loaded and played
-            // TODO there is a bug outstanding where stings are appearing as mainContent
             SMILMetadata = event.target.getMetadata(PlayerConstants.SMIL_METADATA_NS);
-
             var contentType:String = SMILMetadata.getValue(PlayerConstants.CONTENT_TYPE);
             switch (contentType) {
                 case PlayerConstants.MAIN_CONTENT_ID :
@@ -269,18 +286,28 @@ public class BatchEventService extends ProxyElement {
         }
     }
 
-    private function AdMetaEvaluation(value:String):void {
+    private function AdMetaEvaluation(value:*):void {
         if (value == AdState.AD_BREAK_COMPLETE) {
             playingMainContent = true;
-            contentViewingSequenceNumber++;
-            eventsManager.addContentEvent(buildAndReturnContentEvent(ContentTypes.MAIN_CONTENT));
-            eventsManager.flushAll();
+            /*contentViewingSequenceNumber++;
+             eventsManager.addContentEvent(buildAndReturnContentEvent(ContentTypes.MAIN_CONTENT));
+             eventsManager.flushAll();*/   //todo this event is triggered from the durationChange, need to check if auditude triggers the durationChange
 
         } else if (value == AdState.AD_BREAK_START) {
             playingMainContent = false;
             contentViewingSequenceNumber++;
-            currentAdBreakSequenceNumber++;
-            eventsManager.addContentEvent(buildAndReturnContentEvent(ContentTypes.AD_BREAK));
+
+        } else if (typeof(value) == "object") {
+            if (value.state == AdState.STARTED) {
+
+                adUrlResource = value.contentUrl;
+                currentAdBreakSequenceNumber++;
+                defineContentUrl(true);
+                eventsManager.addContentEvent(buildAndReturnContentEvent(ContentTypes.AD_BREAK));
+
+            } else if (value.state == AdState.STOPPED) {
+
+            }
 
         } else if (value == AdMetadata.CLICK_THRU) {
             eventsManager.addUserEvent(buildAndReturnUserEvent(UserEventTypes.CLICK));
@@ -305,7 +332,7 @@ public class BatchEventService extends ProxyElement {
             if (event.value == "playing") {
                 userEventType = UserEventTypes.PLAY;
             }
-            if (event.value == "pause") {
+            if (event.value == "paused") {
                 userEventType = UserEventTypes.PAUSE;
             }
         }
@@ -354,33 +381,8 @@ public class BatchEventService extends ProxyElement {
             case MediaTraitType.DYNAMIC_STREAM:
                 toggleDynamicStreamListeners(added);
                 break;
-            case MediaTraitType.DISPLAY_OBJECT:
-                toggleDisplayListeners(added);
-                break;
         }
     }
-
-
-    private function toggleDisplayListeners(added:Boolean):void {
-        var display:DisplayObjectTrait = proxiedElement.getTrait(MediaTraitType.DISPLAY_OBJECT) as DisplayObjectTrait;
-
-        if (display) {
-            display.addEventListener(DisplayObjectEvent.DISPLAY_OBJECT_CHANGE, onDisplayObjectChange);
-            display.addEventListener(DisplayObjectEvent.MEDIA_SIZE_CHANGE, onMediaSizeChange);
-        } else {
-            display.removeEventListener(DisplayObjectEvent.DISPLAY_OBJECT_CHANGE, onDisplayObjectChange);
-            display.removeEventListener(DisplayObjectEvent.MEDIA_SIZE_CHANGE, onMediaSizeChange);
-        }
-    }
-
-    private function onMediaSizeChange(event:DisplayObjectEvent):void {
-        logger.debug("On Media Size Change old:{0}x{1} new:{2}x{3}", event.oldHeight, event.oldWidth, event.newHeight, event.newWidth);
-    }
-
-    private function onDisplayObjectChange(event:DisplayObjectEvent):void {
-        logger.debug("On Display Object Change old:{0} new:{1}", event.oldDisplayObject, event.newDisplayObject)
-    }
-
 
     private function togglePlayListeners(added:Boolean):void {
         playable = proxiedElement.getTrait(MediaTraitType.PLAY) as PlayTrait;
@@ -395,7 +397,9 @@ public class BatchEventService extends ProxyElement {
     }
 
     private function toggleDynamicStreamListeners(added:Boolean):void {
+
         dynamicStream = proxiedElement.getTrait(MediaTraitType.DYNAMIC_STREAM) as DynamicStreamTrait;
+
         if (dynamicStream) {
             if (added) {
                 dynamicStream.addEventListener(DynamicStreamEvent.AUTO_SWITCH_CHANGE, onAutoSwitchChange);
@@ -409,15 +413,6 @@ public class BatchEventService extends ProxyElement {
 
     private function toggleLoadListeners(added:Boolean):void {
         loadable = proxiedElement.getTrait(MediaTraitType.LOAD) as LoadTrait;
-        if (loadable) {
-
-
-            trace(loadable.loadState);
-            /* loadable.addEventListener(LoadEvent.LOAD_STATE_CHANGE, onLoadableStateChange);
-             loadable.addEventListener(LoadEvent.BYTES_TOTAL_CHANGE, onBytesTotalChange);*/
-
-
-        }
     }
 
     private function onAutoSwitchChange(event:DynamicStreamEvent):void {
@@ -431,28 +426,23 @@ public class BatchEventService extends ProxyElement {
         }
     }
 
-    private function onBytesTotalChange(event:LoadEvent):void {
-    }
-
-    private function onLoadableStateChange(event:LoadEvent):void {
-        trace(event);
-    }
-
-
     private function onPlayStateChange(event:PlayEvent):void {
         if (playingMainContent) {
             switch (event.playState) {
                 case PlayState.PAUSED:
                     cumulativeDurationMonitor.stop();
+                    cumulativeDurationFlushTimer.stop();
                     break;
                 case PlayState.PLAYING:
                     if (!cumulativeDurationMonitor.running) {
                         cumulativeDurationMonitor.start();
+                        cumulativeDurationFlushTimer.start();
                     }
                     break;
                 case PlayState.STOPPED:
                     if (cumulativeDurationMonitor.running) {
                         cumulativeDurationMonitor.stop();
+                        cumulativeDurationFlushTimer.stop();
                     }
                     break;
             }
@@ -468,8 +458,8 @@ public class BatchEventService extends ProxyElement {
     }
 
     private function onTraitRemove(event:MediaElementEvent):void {
-       /// processTrait(event.traitType, false);
-         var traitType:String;
+        /// processTrait(event.traitType, false);
+        var traitType:String;
         for each (traitType in event.target.traitTypes) {
             processTrait(traitType, false);
         }
@@ -480,8 +470,7 @@ public class BatchEventService extends ProxyElement {
         if (buffer) {
             if (added) {
                 buffer.addEventListener(BufferEvent.BUFFERING_CHANGE, onBufferingChange);
-            }
-            else {
+            } else {
                 buffer.removeEventListener(BufferEvent.BUFFERING_CHANGE, onBufferingChange);
             }
         }
@@ -501,7 +490,10 @@ public class BatchEventService extends ProxyElement {
         if (playingMainContent) {
             if (event.seeking) {
                 if (!seeking) {
-                    cumulativeDurationMonitor.stop();
+                    cumulativeDurationMonitor.stop(); //todo check this checker is actually working accurately....
+
+                    contentViewingSequenceNumber = evaluateContentViewingSeqNum();
+
                     eventsManager.addUserEvent(buildAndReturnUserEvent(UserEventTypes.SCRUB));
                     seeking = event.seeking;
                 }
@@ -536,13 +528,22 @@ public class BatchEventService extends ProxyElement {
     private function onDurationChange(event:TimeEvent):void {
         ///MetaDataAdded of the MediaElementEvent, fires 3 times, so wiring this to the duration change seems to be the next most reliable event....
         if (playingMainContent) {
+
+            if (!cumulativeDurationMonitor.running) cumulativeDurationMonitor.start();    // this should only fire when the main content starts...
+
+            contentUrl = "mainResource";
             defineContentUrl(false);
-            contentViewingSequenceNumber++;
+
+            contentViewingSequenceNumber++;        // content sequence has changed by 1. same occurs when an advert starts
+            currentAdBreakSequenceNumber = 0;     ///we are not in an adBreak so set it to 0
             mainContentCount++;
+
             eventsManager.addContentEvent(buildAndReturnContentEvent(ContentTypes.MAIN_CONTENT));
             eventsManager.flushAll();
         } else if (!adMetadata) {
             defineContentUrl(true);
+
+            contentViewingSequenceNumber++;
             eventsManager.addContentEvent(buildAndReturnContentEvent(ContentTypes.AD_BREAK));
         }
     }
@@ -552,12 +553,21 @@ public class BatchEventService extends ProxyElement {
             var streamingUrlResource:StreamingURLResource = loadable.resource as StreamingURLResource;
             if (streamingUrlResource) {
                 contentUrl = streamingUrlResource.url;
+            } else if (adUrlResource) {
+                contentUrl = adUrlResource;
             }
-        } else if (!adMetadata) {
-            contentUrl = "mainResource";
-        } else {
-
         }
+    }
+
+    private function exitEvent():void {
+        eventsManager.addUserEvent(buildAndReturnUserEvent(UserEventTypes.EXIT));
+    }
+
+    private function evaluateContentViewingSeqNum():int {
+        //// TODO match the sequence number against the seekPosition using the adMap.. this could be done in the scrubPreventionProxy  -  NICE TO HAVE
+        var value:int;
+
+        return value;
     }
 
     private function incrementAndGetUserEventId():int {
@@ -566,16 +576,23 @@ public class BatchEventService extends ProxyElement {
     }
 
     private function incrementAndGetContentEventId():int {
-        contentEventId++;
+        if (oldUserEventId == userEventId) {
+            contentEventId++;
+        } else {
+            contentEventId = 0;
+        }
+        oldUserEventId = userEventId;
+
         return contentEventId;
     }
 
     private function buildAndReturnUserEvent(userEventType:String):UserEvent {
+        //TODO generateAssociatedContentEvent(); - there need to be an associated contentEvent.
         return new UserEvent(incrementAndGetUserEventId(), cumulativeDurationCount, userEventType, programmeId);
     }
 
     private function buildAndReturnContentEvent(contentType:String):ContentEvent {
-        return new ContentEvent(isPopupInteractive, mainAssetId, new Date(), isOverlayInteractive, contentViewingSequenceNumber, incrementAndGetContentEventId(), campaignId, cumulativeDurationCount, userEventId, contentDuration, contentType, currentAdBreakSequenceNumber, contentUrl);
+        return new ContentEvent(isPopupInteractive, mainAssetId, new Date(), isOverlayInteractive, contentViewingSequenceNumber, incrementAndGetContentEventId(), campaignId, cumulativeDurationCount, userEventId, cumulativeDurationCount, contentType, currentAdBreakSequenceNumber, contentUrl);
     }
 }
 }
